refactor : Answers로 메소드 이동 및 테스트 코드 작성
