+++ src/test/java/qna/domain/AnswersTest.java
@@ -1,0 +1,21 @@
+package qna.domain;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static qna.domain.AnswerTest.A2;
+import static qna.domain.QuestionTest.*;
+
+public class AnswersTest {
+
+    @Test
+    void deleteAnswers() {
+        DeleteHistories deleteHistories = new DeleteHistories();
+        Answers answers = Q1.getAnswers();
+
+        Q1.addAnswer(A2);
+        answers.delete(deleteHistories);
+        assertThat(A2.isDeleted()).isTrue();
+    }
+
+}
