+++ src/test/java/qna/service/QnaServiceTest.java
@@ -38,6 +38,8 @@ public class QnaServiceTest {
     public void setUp() throws Exception {
         question = new Question(1L, "title1", "contents1").writeBy(UserTest.JAVAJIGI);
         answer = new Answer(11L, UserTest.JAVAJIGI, QuestionTest.Q1, "Answers Contents1");
+        answers = new Answers();
+
         question.addAnswer(answer);
     }
